+++ src/indexable/mod.rs
@@ -1,0 +1,9 @@
+//! # Indexable `Dataloader`.
+
+mod dataloader;
+mod dataset;
+mod fetch;
+pub mod sampler;
+
+pub use dataloader::DataLoader;
+pub use dataset::{Dataset, GetSample, Len, NdarrayDataset};