+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- more strict compiler warnings
+- `Dataloader` for iterable datasets. A dalaloader working for any type that implements `IntoIterator`
+- project folder reorganization
+- default collate now support reference
+
 ## [0.2.1] - 2022-27-09
-### Flixed
+### Fixed
 - Fix link in the doc
 - Add missing documentation field in package metadata
 ## [0.2.0] - 2022-27-09